refactor(Reports): move component to its own directory
